rename module name go-elektra
