+++ src/bindings/go-elektra/kdb/kdb_test.go
@@ -4,8 +4,8 @@ import (
 	"errors"
 	"testing"
 
-	elektra "go.libelektra.org/kdb"
-	. "go.libelektra.org/test"
+	elektra "go.libelektra.org/src/bindings/go-elektra/kdb"
+	. "go.libelektra.org/src/bindings/go-elektra/test"
 )
 
 func TestOpen(t *testing.T) {
