+++ src/bindings/go-elektra/kdb/keyset_test.go
@@ -3,8 +3,8 @@ package kdb_test
 import (
 	"testing"
 
-	elektra "go.libelektra.org/kdb"
-	. "go.libelektra.org/test"
+	elektra "go.libelektra.org/src/bindings/go-elektra/kdb"
+	. "go.libelektra.org/src/bindings/go-elektra/test"
 )
 
 func TestCreateKeySet(t *testing.T) {
